pyproject.toml
[project]
name = "kuberef"
version = "0.1.1"
description = "A Cloud-Native validator for Kubernetes Secret references."
authors = [
    {name = "hudazaan", email = "naazhuda2000@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kubernetes>=29.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "PyYAML>=6.0.1"
]

[tool.poetry]
# This tells Poetry where my code lives
packages = [{include = "kuberef", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
kuberef = "kuberef.main:start"
[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)"
]
